Update CORS config to add beta frontend URL
Added 'https://skillbridge-fe-beta.vercel.app' to the list of allowed origins in the CORS configuration to support requests from the beta frontend. Also, there appears to be a syntax error introduced with an extra quote and parenthesis in the allowed origins list and an extra string in the source initialization.